+++ dune/pdelab/backend/solver.hh
@@ -21,7 +21,7 @@ namespace Dune {
       template<class V>
       typename Dune::template FieldTraits<typename V::ElementType >::real_type norm(const V& v) const
       {
-        return Backend::native(v).two_norm();
+        return v.two_norm();
       }
     };